requirements-local-gpu.txt
numpy==1.14.5
tensorflow-gpu==1.10.0
hyperdash==0.15.3
scipy==1.1.0
pandas==0.22.0
dill==0.2.9
Pillow==5.3.0